Create initial rate checker unit test spec
